Move BaseGenerator logic to core
