+++ csrc/core/BaseGeneratorImpl.cpp
