+++ csrc/core/BaseGeneratorImpl.h
